+++ Model/Processor.php
@@ -3,65 +3,146 @@
 namespace Ingenico\Payment\Model;
 
 use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\DB\TransactionFactory;
+use Magento\Sales\Helper\Data as SalesData;
+use Magento\Sales\Model\Order\Invoice;
+use Magento\Framework\Registry;
+use Ingenico\Payment\Model\Config as IngenicoConfig;
+use Ingenico\Payment\Model\Connector as IngenicoConnector;
+use Magento\Sales\Model\OrderFactory;
+use Magento\Sales\Api\InvoiceRepositoryInterface;
+use Magento\Sales\Api\OrderRepositoryInterface;
+use Magento\Sales\Api\CreditmemoManagementInterface;
+use Magento\Sales\Model\Order\CreditmemoFactory;
+use Magento\Sales\Model\Service\InvoiceService;
+use Magento\Sales\Model\Order\Email\Sender\InvoiceSender;
+use Magento\Sales\Model\Order\Email\Sender\CreditmemoSender;
+use Magento\Sales\Model\Order\Payment\Transaction\BuilderInterface as TransactionBuilder;
+use Magento\Sales\Api\CreditmemoRepositoryInterface;
 
 class Processor
 {
-    protected $_orderFactory;
-    protected $_invoiceRepository;
-    protected $_orderRepository;
-    protected $_creditmemoManagement;
-    protected $_creditmemoFactory;
-    protected $_invoiceService;
-    protected $_invoiceSender;
-    protected $_creditmemoSender;
-    protected $_transactionBuilder;
-    protected $_creditmemoRepository;
-    protected $_registry;
+    /**
+     * @var OrderFactory
+     */
+    private $orderFactory;
 
     /**
-     * @var Config
+     * @var InvoiceRepositoryInterface
+     */
+    private $invoiceRepository;
+
+    /**
+     * @var OrderRepositoryInterface
+     */
+    private $orderRepository;
+
+    /**
+     * @var CreditmemoManagementInterface
+     */
+    private $creditmemoManagement;
+
+    /**
+     * @var CreditmemoFactory
+     */
+    private $creditmemoFactory;
+
+    /**
+     * @var InvoiceService
+     */
+    private $invoiceService;
+
+    /**
+     * @var InvoiceSender
+     */
+    private $invoiceSender;
+
+    /**
+     * @var CreditmemoSender
+     */
+    private $creditmemoSender;
+
+    /**
+     * @var TransactionBuilder
+     */
+    private $transactionBuilder;
+
+    /**
+     * @var CreditmemoRepositoryInterface
+     */
+    private $creditmemoRepository;
+
+    /**
+     * @var Registry
+     */
+    private $registry;
+
+    /**
+     * @var TransactionFactory
+     */
+    private $transactionFactory;
+
+    /**
+     * @var SalesData
+     */
+    private $salesData;
+
+    /**
+     * @var IngenicoConfig
      */
     private $config;
 
     /**
-     * @var \Ingenico\Payment\Model\Connector
+     * @var IngenicoConnector
      */
-    protected $_connector;
+    private $connector;
 
     /**
      * Constructor
      */
     public function __construct(
-        \Magento\Sales\Model\OrderFactory $orderFactory,
-        \Magento\Sales\Api\InvoiceRepositoryInterface $invoiceRepository,
-        \Magento\Sales\Api\OrderRepositoryInterface $orderRepository,
-        \Magento\Sales\Api\CreditmemoManagementInterface $creditmemoManagement,
-        \Magento\Sales\Model\Order\CreditmemoFactory $creditmemoFactory,
-        \Magento\Sales\Model\Service\InvoiceService $invoiceService,
-        \Magento\Sales\Model\Order\Email\Sender\InvoiceSender $invoiceSender,
-        \Magento\Sales\Model\Order\Email\Sender\CreditmemoSender $creditmemoSender,
-        \Magento\Sales\Model\Order\Payment\Transaction\BuilderInterface $transactionBuilder,
-        \Magento\Sales\Api\CreditmemoRepositoryInterface $creditmemoRepository,
-        \Magento\Framework\Registry $registry,
-        \Ingenico\Payment\Model\Config $config
+        OrderFactory $orderFactory,
+        InvoiceRepositoryInterface $invoiceRepository,
+        OrderRepositoryInterface $orderRepository,
+        CreditmemoManagementInterface $creditmemoManagement,
+        CreditmemoFactory $creditmemoFactory,
+        InvoiceService $invoiceService,
+        InvoiceSender $invoiceSender,
+        CreditmemoSender $creditmemoSender,
+        TransactionBuilder $transactionBuilder,
+        CreditmemoRepositoryInterface $creditmemoRepository,
+        TransactionFactory $transactionFactory,
+        SalesData $salesData,
+        Registry $registry,
+        IngenicoConfig $config
     ) {
-        $this->_orderFactory = $orderFactory;
-        $this->_invoiceRepository = $invoiceRepository;
-        $this->_orderRepository = $orderRepository;
-        $this->_creditmemoManagement = $creditmemoManagement;
-        $this->_creditmemoFactory = $creditmemoFactory;
-        $this->_creditmemoRepository = $creditmemoRepository;
-        $this->_invoiceService = $invoiceService;
-        $this->_invoiceSender = $invoiceSender;
-        $this->_creditmemoSender = $creditmemoSender;
-        $this->_transactionBuilder = $transactionBuilder;
-        $this->_registry = $registry;
+        $this->orderFactory      = $orderFactory;
+        $this->invoiceRepository = $invoiceRepository;
+        $this->orderRepository  = $orderRepository;
+        $this->creditmemoManagement = $creditmemoManagement;
+        $this->creditmemoFactory = $creditmemoFactory;
+        $this->creditmemoRepository = $creditmemoRepository;
+        $this->invoiceService = $invoiceService;
+        $this->invoiceSender = $invoiceSender;
+        $this->creditmemoSender = $creditmemoSender;
+        $this->transactionBuilder = $transactionBuilder;
+        $this->transactionFactory = $transactionFactory;
+        $this->salesData = $salesData;
+        $this->registry = $registry;
         $this->config = $config;
     }
 
-    public function setConnector(\Ingenico\Payment\Model\Connector $connector)
+    /**
+     * Set Connector.
+     *
+     * @param IngenicoConnector $connector
+     *
+     * @return $this
+     */
+    public function setConnector(IngenicoConnector $connector)
     {
-        $this->_connector = $connector;
+        $this->connector = $connector;
+
         return $this;
     }
 
@@ -76,11 +157,12 @@ class Processor
     public function getOrderByIncrementId($incrementId)
     {
         if (!$incrementId) {
-            throw new \Magento\Framework\Exception\LocalizedException(__('ingenico.exception.message8'));
+            throw new LocalizedException(__('ingenico.exception.message8'));
         }
-        $order = $this->_orderFactory->create()->loadByIncrementId($incrementId);
+
+        $order = $this->orderFactory->create()->loadByIncrementId($incrementId);
         if (!$order->getId()) {
-            throw new \Magento\Framework\Exception\LocalizedException(__('ingenico.exception.message9', $incrementId));
+            throw new LocalizedException(__('ingenico.exception.message9', $incrementId));
         }
 
         return $order;
@@ -108,19 +190,25 @@ class Processor
 
         // accept only authorizations for new (pending) orders
         if ($order->getStatus() !== $order->getConfig()->getStateDefaultStatus($order::STATE_NEW)) {
-            $this->_addOrderMessage($order, __('ingenico.notification.message2', $authorizedStatus, $order->getStatus()));
-            return $this->_orderRepository->save($order);
+            $this->_addOrderMessage(
+                $order,
+                __('ingenico.notification.message2', $authorizedStatus, $order->getStatus())
+            );
+
+            return $this->orderRepository->save($order);
         }
 
         // Set order status
         $new_status = $this->config->getOrderStatusAuth();
         $status = $this->config->getAssignedState($new_status);
+
         $order->setData('state', $status->getState());
         $order->setStatus($status->getStatus());
-        $this->_addOrderMessage($order, $message, __('ingenico.notification.message3'));
-        $this->_registry->register($this->_connector::REGISTRY_KEY_CAN_SEND_AUTH_EMAIL, true, true);
 
-        return $this->_orderRepository->save($order);
+        $this->_addOrderMessage($order, $message, __('ingenico.notification.message3'));
+        $this->registry->register($this->connector::REGISTRY_KEY_CAN_SEND_AUTH_EMAIL, true, true);
+
+        return $this->orderRepository->save($order);
     }
 
     /**
@@ -138,13 +226,14 @@ class Processor
         $order = $this->getOrderByIncrementId($incrementId);
 
         // only process if request is not from Admin Panel
-        if ($this->_registry->registry('current_invoice')) {
+        if ($this->registry->registry('current_invoice')) {
             return $order;
         }
 
         if ($order->isCanceled()) {
             $this->_addOrderMessage($order, __('ingenico.notification.message5'));
-            return $this->_orderRepository->save($order);
+
+            return $this->orderRepository->save($order);
         }
 
         $processStatus = false;
@@ -152,36 +241,47 @@ class Processor
         $trxId = $paymentResult->getPayId() . '-' . $paymentResult->getPayIdSub();
         if ($order->hasInvoices()) {
             foreach ($order->getInvoiceCollection() as $invoice) {
+                /** @var Invoice $invoice */
                 if ($invoice->getTransactionId() == $trxId && $invoice->canCancel()) {
                     $invoice->pay();
-                    $this->_invoiceRepository->save($invoice);
-                    $this->_orderRepository->save($invoice->getOrder());
-                    if (!$invoice->getEmailSent()) {
-                        $this->_invoiceSender->send($invoice);
+                    $this->invoiceRepository->save($invoice);
+                    $this->orderRepository->save($invoice->getOrder());
+
+                    if (!$invoice->getEmailSent() && $this->salesData->canSendNewInvoiceEmail()) {
+                        $this->invoiceSender->send($invoice);
                     }
+
                     $processStatus = true;
                 }
             }
-
         } else {
             try {
-                $invoice = $this->_invoiceService->prepareInvoice($order);
+                $invoice = $this->invoiceService->prepareInvoice($order);
                 $invoice->setRequestedCaptureCase($invoice::CAPTURE_ONLINE);
                 $invoice->register();
                 $invoice->getOrder()->setIsInProcess(true);
                 $invoice->setIsPaid(true);
+
                 if ($paymentResult->getPm() !== 'Bank transfer') {
                     $invoice->setTransactionId($trxId);
                 }
-                $this->_invoiceRepository->save($invoice);
-                $this->_invoiceSender->send($invoice);
-                $processStatus = true;
 
-            } catch (\Exception $e) {
-                $this->_connector->log($e->getMessage(), 'crit');
+                $dbTransaction = $this->transactionFactory->create();
+                $dbTransaction->addObject($invoice)
+                              ->addObject($invoice->getOrder())
+                              ->save();
+
+                if ($this->salesData->canSendNewInvoiceEmail()) {
+                    $this->invoiceSender->send($invoice);
+                }
+
+                $processStatus = true;
             } catch (LocalizedException $e) {
-                $this->_connector->log(sprintf('%s::%s %s', __CLASS__, __METHOD__, $e->getMessage()));
+                $this->connector->log(sprintf('%s::%s %s', __CLASS__, __METHOD__, $e->getMessage()));
+
                 throw $e;
+            } catch (\Exception $e) {
+                $this->connector->log($e->getMessage(), 'crit');
             }
         }
 
@@ -194,7 +294,7 @@ class Processor
             $this->_addOrderMessage($order, $message, __('ingenico.notification.message6'));
         }
 
-        return $this->_orderRepository->save($order);
+        return $this->orderRepository->save($order);
     }
 
     /**
@@ -220,7 +320,8 @@ class Processor
     {
         $order = $this->getOrderByIncrementId($incrementId);
         $this->_addOrderMessage($order, $message);
-        return $this->_orderRepository->save($order);
+
+        return $this->orderRepository->save($order);
     }
 
     /**
@@ -239,7 +340,8 @@ class Processor
 
         if ($order->isCanceled()) {
             $this->_addOrderMessage($order, __('ingenico.notification.message5'));
-            return $this->_orderRepository->save($order);
+
+            return $this->orderRepository->save($order);
         }
 
         try {
@@ -251,8 +353,8 @@ class Processor
                 if ($creditMemo->getTransactionId() === $trxId && $creditMemo->canRefund()) {
                     $creditMemo->setPaymentRefundDisallowed(true);
                     // @see \Magento\Sales\Model\Service\CreditmemoService::refund()
-                    $this->_creditmemoManagement->refund($creditMemo, false);
-                    $this->_creditmemoSender->send($creditMemo);
+                    $this->creditmemoManagement->refund($creditMemo, false);
+                    $this->creditmemoSender->send($creditMemo);
 
                     $hasFound = true;
                 }
@@ -272,26 +374,26 @@ class Processor
                 }
 
                 /** @var \Magento\Sales\Model\Order\Creditmemo $creditMemo */
-                $creditMemo = $this->_creditmemoFactory->createByOrder($order);
+                $creditMemo = $this->creditmemoFactory->createByOrder($order);
                 $creditMemo->setTransactionId($trxId)
                     ->setInvoice($invoiceToRefund)
                     ->setBaseGrandTotal($paymentResult->getAmount())
                     ->setGrandTotal($paymentResult->getAmount())
                     ->setPaymentRefundDisallowed(true);
 
-                $this->_creditmemoManagement->refund($creditMemo);
-                $this->_creditmemoSender->send($creditMemo);
+                $this->creditmemoManagement->refund($creditMemo);
+                $this->creditmemoSender->send($creditMemo);
             }
 
-            $order = $this->_orderRepository->get($order->getId());
+            $order = $this->orderRepository->get($order->getId());
             $this->_addOrderMessage($order, $message);
         } catch (LocalizedException $e) {
-            $this->_connector->log(sprintf('%s::%s %s', __CLASS__, __METHOD__, $e->getMessage()));
+            $this->connector->log(sprintf('%s::%s %s', __CLASS__, __METHOD__, $e->getMessage()));
         } catch (\Exception $e) {
-            $this->_connector->log($e->getMessage(), 'crit');
+            $this->connector->log($e->getMessage(), 'crit');
         }
 
-        return $this->_orderRepository->save($order);
+        return $this->orderRepository->save($order);
     }
 
     /**
@@ -310,14 +412,14 @@ class Processor
 
         if (!$order->canCancel()) {
             $this->_addOrderMessage($order, __('ingenico.notification.message7', $message));
-            return $this->_orderRepository->save($order);
+
+            return $this->orderRepository->save($order);
         }
 
         $order->cancel();
-
         $this->_addOrderMessage($order, $message, __('ingenico.notification.message8'));
 
-        return $this->_orderRepository->save($order);
+        return $this->orderRepository->save($order);
     }
 
     protected function _addOrderMessage($order, $message, $fallbackMsg = null)