+++ Plugin/MagentoSalesModelOrder.php
@@ -47,10 +47,12 @@ class MagentoSalesModelOrder
      */
     public function afterGetCanSendNewEmailFlag(\Magento\Sales\Model\Order $subject, $result)
     {
-        if (in_array($subject->getPayment()->getMethod(),
+        if (in_array(
+            $subject->getPayment()->getMethod(),
             array_merge($this->ingenicoHelper->getPaymentMethodCodes(), [
                 \Ingenico\Payment\Model\Method\Alias::PAYMENT_METHOD_CODE
-            ]))
+            ])
+        )
         ) {
             if ($this->cnf->getOrderConfirmationEmailMode() === OrderEmail::STATUS_ENABLED) {
                 return $result;